requirements.txt
PyQt5==5.15.7
sqlalchemy==1.4.44
psycopg2-binary==2.9.5
pyyaml==6.0
nanoid==2.0.0
pydantic==1.10.2
# qt beauty
qdarkstyle==3.1
typer==0.7.0
# encryption
cryptography==38.0.4
rsa==4.9
pymdown-extensions==9.9
markdown2[all]==2.4.5